Initial support for Capella testing
