+++ Cargo.toml
@@ -1,3 +1,4 @@
+workspace = { members = ["utilities"] }
 [package]
 name = "couchbase-shell"
 version = "0.95.0"
@@ -55,8 +56,8 @@ tokio = { version = "1.38.0", features = ["full"] }
 tokio-util = { version = "0.7.11", features = ["codec"] }
 tokio-rustls = { version = "0.26.0" }
 toml = "0.8.14"
-trust-dns-resolver = { version = "0.23.2", features = ["dns-over-rustls"] }
 uuid = { version = "1.9.1", features = ["v4"] }
+utilities = { path = "utilities" }
 webpki-roots = "0.26.3"
 
 [target.'cfg(not(any(target_os = "macos", target_os = "windows")))'.dependencies]
@@ -83,14 +84,16 @@ static-link-openssl = ["dep:openssl"]
 # We use a list of supported features to determine which tests we want to run, the defaults will always be run, then we
 # define any additional features using the specific server versions we test against
 default = ["key_value", "query", "query_index", "query_index_definitions", "query_index_advise"]
-"7.1.1" = ["collections"]
-"7.0.3" = ["collections"]
+"7.1.1" = ["collections", "import_sample"]
+"7.0.3" = ["collections", "import_sample"]
+"capella" = ["collections"]
 key_value = []
 query = []
 collections = []
 query_index = []
 query_index_definitions = []
 query_index_advise = []
+import_sample = []
 
 [[bin]]
 name = "cbsh"
