+++ src/client/tls.rs
@@ -4,70 +4,12 @@ use crate::ClusterTlsConfig;
 use log::{debug, error};
 use rustls_pemfile::{read_all, Item};
 use std::convert::TryFrom;
-use std::fmt::Debug;
 use std::fs;
 use std::io::BufReader;
 use std::sync::Arc;
-use tokio_rustls::rustls::client::danger::{
-    HandshakeSignatureValid, ServerCertVerified, ServerCertVerifier,
-};
 use tokio_rustls::rustls::crypto::{aws_lc_rs::default_provider, CryptoProvider};
-use tokio_rustls::rustls::pki_types::{CertificateDer, ServerName, UnixTime};
-use tokio_rustls::rustls::{
-    ClientConfig, DigitallySignedStruct, Error, RootCertStore, SignatureScheme,
-};
-
-#[derive(Debug)]
-struct InsecureCertVerifier {}
-
-impl ServerCertVerifier for InsecureCertVerifier {
-    fn verify_server_cert(
-        &self,
-        _end_entity: &CertificateDer,
-        _intermediates: &[CertificateDer],
-        _server_name: &ServerName,
-        _ocsp_response: &[u8],
-        _now: UnixTime,
-    ) -> Result<ServerCertVerified, Error> {
-        Ok(ServerCertVerified::assertion())
-    }
-
-    fn verify_tls12_signature(
-        &self,
-        _message: &[u8],
-        _cert: &CertificateDer<'_>,
-        _dss: &DigitallySignedStruct,
-    ) -> Result<HandshakeSignatureValid, Error> {
-        Ok(HandshakeSignatureValid::assertion())
-    }
-
-    fn verify_tls13_signature(
-        &self,
-        _message: &[u8],
-        _cert: &CertificateDer<'_>,
-        _dss: &DigitallySignedStruct,
-    ) -> Result<HandshakeSignatureValid, Error> {
-        Ok(HandshakeSignatureValid::assertion())
-    }
-
-    fn supported_verify_schemes(&self) -> Vec<SignatureScheme> {
-        vec![
-            SignatureScheme::RSA_PKCS1_SHA1,
-            SignatureScheme::ECDSA_SHA1_Legacy,
-            SignatureScheme::RSA_PKCS1_SHA256,
-            SignatureScheme::ECDSA_NISTP256_SHA256,
-            SignatureScheme::RSA_PKCS1_SHA384,
-            SignatureScheme::ECDSA_NISTP384_SHA384,
-            SignatureScheme::RSA_PKCS1_SHA512,
-            SignatureScheme::ECDSA_NISTP521_SHA512,
-            SignatureScheme::RSA_PSS_SHA256,
-            SignatureScheme::RSA_PSS_SHA384,
-            SignatureScheme::RSA_PSS_SHA512,
-            SignatureScheme::ED25519,
-            SignatureScheme::ED448,
-        ]
-    }
-}
+use tokio_rustls::rustls::{ClientConfig, RootCertStore};
+use utilities;
 
 #[derive(Clone)]
 pub struct RustTlsConfig {
@@ -87,7 +29,7 @@ impl RustTlsConfig {
         if accept_all_certs {
             let config = builder
                 .dangerous()
-                .with_custom_certificate_verifier(Arc::new(InsecureCertVerifier {}))
+                .with_custom_certificate_verifier(Arc::new(utilities::InsecureCertVerifier {}))
                 .with_no_client_auth();
 
             return Ok(RustTlsConfig {
