+++ utilities/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "utilities"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tokio-rustls = "0.26.0"
+trust-dns-resolver = "0.23.2"
+
